+++ src/lib/libdpd/buf4_dot.cc
@@ -39,7 +39,7 @@ double dpd_buf4_dot(dpdbuf4 *BufA, dpdbuf4 *BufB)
       if(!rows_per_bucket)
 	dpd_error("buf4_dot: Not enough memory for one row!", stderr);
 
-      nbuckets = ceil((double) BufA->params->rowtot[h]/
+      nbuckets = (int) ceil((double) BufA->params->rowtot[h]/
 		      (double) rows_per_bucket);
 
       rows_left = BufA->params->rowtot[h] % rows_per_bucket;
@@ -102,3 +102,4 @@ double dpd_buf4_dot(dpdbuf4 *BufA, dpdbuf4 *BufB)
 }
       
 } /* extern "C" */
+
